+++ processor/src/test/resources/expected/com/example/repository/Arez_CompleteRepositoryExample.java
@@ -128,6 +128,7 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setName' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }
+    assert null != name;
     if ( !Objects.equals( name, super.getName() ) ) {
       this.$$arez$$_name.preReportChanged();
       super.setName(name);
@@ -150,6 +151,7 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setPackageName' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }
+    assert null != packageName;
     if ( !Objects.equals( packageName, super.getPackageName() ) ) {
       this.$$arez$$_packageName.preReportChanged();
       super.setPackageName(packageName);