+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Update the 'org.realityforge.arez' dependencies to version '0.88'
+
 ### [v0.13](https://github.com/arez/arez-browserlocation/tree/v0.13) (2018-05-28)
 [Full Changelog](https://github.com/arez/arez-browserlocation/compare/v0.12...v0.13)